lambda_files/requirements.txt
PyGithub
pandas
numpy
PyMySQL
python-dotenv